@@ -1,178 +1,115 @@
 package art
 
 import (
-	"bytes"
+	"flag"
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestTreeDump4(t *testing.T) {
-	n4 := factory.newNode4()
-	n4leaf := factory.newLeaf([]byte("key4"), "value4")
-	n4._addChild4('k', true, n4leaf)
+var updateGolden = flag.Bool("update-golden", false, "update .golden files")
 
-	tree := &tree{root: n4}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node4")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [107 · · ·] [k···]")
-	assert.Contains(t, o, "children(1): [0x")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(4): [107 101 121 52] [key4]")
-	assert.Contains(t, o, "val: value4")
-}
-
-func TestTreeDump4BinaryValue(t *testing.T) {
-	n4 := factory.newNode4()
-	n4leaf := factory.newLeaf([]byte("key4"), []byte("value4"))
-	n4._addChild4('k', true, n4leaf)
-
-	tree := &tree{root: n4}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node4")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [107 · · ·] [k···]")
-	assert.Contains(t, o, "children(1): [0x")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(4): [107 101 121 52] [key4]")
-	assert.Contains(t, o, "val: value4", "Binary value should be print as a string")
-}
-
-func TestTreeDump4Int(t *testing.T) {
-	n4 := factory.newNode4()
-	n4leaf := factory.newLeaf([]byte("key4"), 4)
-	n4._addChild4('k', true, n4leaf)
-
-	tree := &tree{root: n4}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node4")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [107 · · ·] [k···]")
-	assert.Contains(t, o, "children(1): [0x")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(4): [107 101 121 52] [key4]")
-	assert.Contains(t, o, "val: 4", "Int value should be print as a number")
-}
-
-func TestTreeDump2NodeWithIntValue(t *testing.T) {
-	n16 := factory.newNode16()
-	n16_2 := factory.newNode16()
-	n16_2leaf := factory.newLeaf([]byte("zima"), 4)
-	n16_2.addChild('z', true, n16_2leaf)
-
-	n16leaf := factory.newLeaf([]byte("key4"), 4)
-	c4leaf := factory.newLeaf([]byte("cey4"), 44)
-	n16.addChild('k', true, n16leaf)
-	n16.addChild('c', true, c4leaf)
-	n16.addChild('z', true, n16_2)
-
-	tree := &tree{root: n16}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node16")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [99 107 122 ")
-	assert.Contains(t, o, "keys: [122 · · · · · · · · · · · · · · ·]")
-	assert.Contains(t, o, "children(3)")
-	assert.Contains(t, o, "children(1)")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(4): [107 101 121 52] [key4]")
-	assert.Contains(t, o, "key(4): [122 105 109 97] [zima]")
-	assert.Contains(t, o, "val: 4", "Int value should be print as a number")
-}
-
-func TestTreeDump16(t *testing.T) {
-	n16 := factory.newNode16()
-	n16leaf := factory.newLeaf([]byte("key16"), "value16")
-	n16.addChild('k', true, n16leaf)
-
-	tree := &tree{root: n16}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node16")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [107 · · · · · · · · · · · · · · ·] [k···············]")
-	assert.Contains(t, o, "children(1)")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(5): [107 101 121 49 54] [key16]")
-	assert.Contains(t, o, "val: value16")
-}
-
-func TestTreeDump48(t *testing.T) {
-	n48 := factory.newNode48()
-	n48leaf := factory.newLeaf([]byte("key48"), "value48")
-	n48.addChild('k', true, n48leaf)
-
-	tree := &tree{root: n48}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node48")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "keys: [· · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · ·  0 · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · · ·]")
-	assert.Contains(t, o, "children(1)")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(5): [107 101 121 52 56] [key48]")
-	assert.Contains(t, o, "val: value48")
-}
-
-func TestTreeDump256(t *testing.T) {
-	n256 := factory.newNode256()
-	n256leaf := factory.newLeaf([]byte("key256"), "value256")
-	n256.addChild('k', true, n256leaf)
-
-	tree := &tree{root: n256}
-	o := tree.String()
-
-	assert.Contains(t, o, "Node256")
-	assert.Contains(t, o, "prefix(0): [0 0 0 0 0 0 0 0 0 0]")
-	assert.Contains(t, o, "children(1)")
-	assert.Contains(t, o, "Leaf")
-	assert.Contains(t, o, "key(6): [107 101 121 50 53 54] [key256]")
-	assert.Contains(t, o, "val: value256")
-}
-
-func TestAppendForExtralOptions(t *testing.T) {
-	tsDec := &treeStringer{make([]depthStorage, 1024), bytes.NewBufferString("")}
-	tsDec.append([]byte{1, 2, 3, 4}, printValuesAsDecimal)
-	assert.Equal(t, "[1 2 3 4]", tsDec.buf.String())
-
-	tsHex := &treeStringer{make([]depthStorage, 1024), bytes.NewBufferString("")}
-	tsHex.appendKey(
-		[]byte{
-			0, // not present
-			1,
-			2,
-			16,
-			17,
+func TestTreeStringer(t *testing.T) {
+	tests := []struct {
+		name   string
+		tree   func() *tree
+		golden string
+	}{
+		{
+			name: "Dump4",
+			tree: func() *tree {
+				n4 := factory.newNode4()
+				n4leaf := factory.newLeaf([]byte("key4"), "value4")
+				n4.addChild('k', true, n4leaf)
+				return &tree{root: n4}
+			},
+			golden: "test/stringer/dump4.golden",
 		},
-		[]byte{
-			0,
-			1,
-			1,
-			1,
-			1,
-		}, printValuesAsHex)
-	assert.Equal(t, "[·  1  2 10 11]", tsHex.buf.String())
-
-	tsDef := &treeStringer{make([]depthStorage, 1024), bytes.NewBufferString("")}
-	tsDef.appendKey(
-		[]byte{
-			0, // not present
-			1,
-			2,
-			16,
-			17,
+		{
+			name: "Dump4BinaryValue",
+			tree: func() *tree {
+				n4 := factory.newNode4()
+				n4leaf := factory.newLeaf([]byte("key4"), []byte("value4"))
+				n4.addChild('k', true, n4leaf)
+				return &tree{root: n4}
+			},
+			golden: "test/stringer/dump4_binary_value.golden",
 		},
-		[]byte{
-			0,
-			1,
-			1,
-			1,
-			1,
-		}, 0)
-	assert.Equal(t, "[·\x01\x02\x10\x11]", tsDef.buf.String())
+		{
+			name: "Dump4Int",
+			tree: func() *tree {
+				n4 := factory.newNode4()
+				n4leaf := factory.newLeaf([]byte("key4"), 4)
+				n4.addChild('k', true, n4leaf)
+				return &tree{root: n4}
+			},
+			golden: "test/stringer/dump4_int.golden",
+		},
+		{
+			name: "Dump16IntValue",
+			tree: func() *tree {
+				n16 := factory.newNode16()
+				n16_2 := factory.newNode16()
+				n16_2leaf := factory.newLeaf([]byte("4yek"), 4)
+				n16_2.addChild('z', true, n16_2leaf)
+
+				n16leaf := factory.newLeaf([]byte("key4"), 4)
+				c4leaf := factory.newLeaf([]byte("cey4"), 44)
+				n16.addChild('k', true, n16leaf)
+				n16.addChild('c', true, c4leaf)
+				n16.addChild('z', true, n16_2)
+				return &tree{root: n16}
+			},
+			golden: "test/stringer/dump16_int_value.golden",
+		},
+		{
+			name: "Dump16",
+			tree: func() *tree {
+				n16 := factory.newNode16()
+				n16leaf := factory.newLeaf([]byte("key16"), "value16")
+				n16.addChild('k', true, n16leaf)
+				return &tree{root: n16}
+			},
+			golden: "test/stringer/dump16.golden",
+		},
+		{
+			name: "Dump48",
+			tree: func() *tree {
+				n48 := factory.newNode48()
+				n48leaf := factory.newLeaf([]byte("key48"), "value48")
+				n48.addChild('k', true, n48leaf)
+				return &tree{root: n48}
+			},
+			golden: "test/stringer/dump48.golden",
+		},
+		{
+			name: "Dump256",
+			tree: func() *tree {
+				n256 := factory.newNode256()
+				n256leaf := factory.newLeaf([]byte("key256"), "value256")
+				n256.addChild('k', true, n256leaf)
+				return &tree{root: n256}
+			},
+			golden: "test/stringer/dump256.golden",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			actualOut := TreeStringer(tt.tree())
+
+			if *updateGolden {
+				t.Logf("%s: updating golden file %s...", tt.name, tt.golden)
+				err := os.WriteFile(tt.golden, []byte(actualOut), 0644)
+				t.Logf("%s: updated golden file %s: err=%v", tt.name, tt.golden, err)
+				require.NoError(t, err)
+			}
+
+			goldenOut, err := os.ReadFile(tt.golden)
+			require.NoError(t, err)
+			assert.Equal(t, string(goldenOut), actualOut)
+		})
+	}
 }
